Замечания + улучшения
