+++ cmd/shortener/main.go
@@ -2,8 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/zauremazhikovayandex/url/cmd/routers"
+	"github.com/zauremazhikovayandex/url/internal/app"
 	"github.com/zauremazhikovayandex/url/internal/config"
+	"github.com/zauremazhikovayandex/url/internal/db/storage"
 	"net/http"
 )
 
@@ -15,7 +16,8 @@ func main() {
 
 func run() error {
 	config.InitConfig()
+	storage.InitStorage()
 
 	fmt.Println("Running server on", config.AppConfig.ServerAddr)
-	return http.ListenAndServe(config.AppConfig.ServerAddr, routers.Router())
+	return http.ListenAndServe(config.AppConfig.ServerAddr, app.Router())
 }
