+++ internal/app/handler.go
@@ -1,1 +1,66 @@
 package app
+
+import (
+	"crypto/rand"
+	"encoding/base64"
+	"fmt"
+	"github.com/go-chi/chi/v5"
+	"github.com/zauremazhikovayandex/url/internal/config"
+	"github.com/zauremazhikovayandex/url/internal/db/storage"
+	"io"
+	"log"
+	"net/http"
+	"strings"
+)
+
+func generateShortID(n int) (string, error) {
+	b := make([]byte, n)
+	_, err := rand.Read(b)
+	if err != nil {
+		return "", err
+	}
+	return base64.URLEncoding.EncodeToString(b)[:n], nil
+}
+
+func PostHandler(w http.ResponseWriter, r *http.Request) {
+
+	body, err := io.ReadAll(r.Body)
+	if err != nil || len(body) == 0 {
+		http.Error(w, "Invalid request body", http.StatusBadRequest)
+		return
+	}
+
+	originalURL := strings.TrimSpace(string(body))
+	id, err := generateShortID(8)
+	if err != nil || id == "" {
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+
+	storage.Store.Set(id, originalURL)
+
+	shortURL := fmt.Sprintf("%s/%s", config.AppConfig.BaseURL, id)
+	w.Header().Set("Content-Type", "text/plain")
+	w.WriteHeader(http.StatusCreated)
+	_, err = w.Write([]byte(shortURL))
+	if err != nil {
+		log.Println(err)
+	}
+
+}
+
+func GetHandler(w http.ResponseWriter, r *http.Request) {
+	id := chi.URLParam(r, "id")
+	if id == "" {
+		http.Error(w, "Missing ID", http.StatusBadRequest)
+		return
+	}
+
+	originalURL, ok := storage.Store.Get(id)
+	if !ok || originalURL == "" {
+		http.Error(w, "URL not found", http.StatusBadRequest)
+		return
+	}
+
+	http.Redirect(w, r, originalURL, http.StatusTemporaryRedirect)
+}
